Separate three-agent proof report responsibilities
The green implementation left report rendering, report writing, and three-agent verification logic in larger modules. This splits those responsibilities so the MVP proof code stays reviewable while preserving the same claim semantics.

Constraint: AGENTS.md requires a mandatory refactor phase with line-budget and single-responsibility checks

Rejected: Leave runner/report/verify oversized | future changes would keep accumulating in broad files

Confidence: high

Scope-risk: narrow

Directive: Keep report generation, report writing, and verifier-specific invariants in separate modules

Tested: cargo fmt --check

Tested: CARGO_TARGET_DIR=target/mvp-demo-proof CARGO_BUILD_JOBS=1 CARGO_INCREMENTAL=0 cargo test -p kamn-e2e-harness --test mvp_demo_claim_contract -- --nocapture

Tested: CARGO_TARGET_DIR=target/mvp-demo-proof CARGO_BUILD_JOBS=1 CARGO_INCREMENTAL=0 cargo test -p kamn-e2e-harness --test mvp_demo_three_agent_claim_contract -- --nocapture

Not-tested: Full make demo-mvp integration
