+++ crates/kamn-e2e-harness/src/mvp_demo/report.rs
@@ -3,7 +3,7 @@ use std::path::Path;
 use super::devnet_settlement::{
     devnet_no_go_reason, devnet_settlement_claim_json, DevnetSettlementEvidence,
 };
-use super::report_artifacts::{artifact_path, artifacts_json};
+use super::report_artifacts::artifacts_json;
 use super::three_agent_claim::three_agent_escrow_claim_json;
 
 /// MVP demo proof report schema marker.
@@ -44,52 +44,7 @@ pub(crate) fn render_report_json(input: &DemoReportInput<'_>) -> String {
     )
 }
 
-pub(crate) fn render_report_markdown(input: &DemoReportInput<'_>) -> String {
-    [
-        markdown_header(input),
-        markdown_artifacts(input),
-        markdown_claim_boundaries().to_owned(),
-    ]
-    .join("\n")
-}
-
-fn markdown_header(input: &DemoReportInput<'_>) -> String {
-    let status = report_status(input);
-    format!(
-        "# KAMN MVP Demo Proof Report\n\n- Run ID: `{}`\n- Status: `{}`\n- Devnet mode: `{}`\n- Report JSON: `{}`\n",
-        input.run_id,
-        status,
-        input.devnet_mode,
-        input.output_root.join("latest/proof/report.json").display(),
-    )
-}
-
-fn markdown_artifacts(input: &DemoReportInput<'_>) -> String {
-    format!(
-        "## Proof Artifacts\n\n- SDK localhost signed artifact: `{}`\n- SDK localhost signed output: `{}`\n- Service API vertical slice output: `{}`\n- Service API websocket output: `{}`\n- Devnet settlement output: `{}`\n- Audit export: `{}`\n",
-        artifact_path(input, &format!("{}/proof/localhost-signed-demo.json", input.run_id)),
-        artifact_path(
-            input,
-            &format!("{}/proof/localhost-signed-demo-output.txt", input.run_id)
-        ),
-        artifact_path(
-            input,
-            &format!(
-                "{}/proof/service-api-vertical-slice-output.txt",
-                input.run_id
-            ),
-        ),
-        artifact_path(input, &format!("{}/proof/service-api-websocket-output.txt", input.run_id)),
-        artifact_path(input, &format!("{}/proof/devnet-settlement-output.txt", input.run_id)),
-        artifact_path(input, &format!("{}/proof/audit-export.json", input.run_id))
-    )
-}
-
-fn markdown_claim_boundaries() -> &'static str {
-    "## Claim Boundaries\n\n- Local runtime, auth, message/task, state, relay, websocket, and audit proof are local-only MVP claims.\n- Settlement or asset movement is not claimed unless the JSON report carries `devnet-backed` evidence.\n- Devnet-required runs without configured settlement evidence are explicit `NO-GO`, not local-only success.\n- Devnet tokens are Solana devnet only and are not real economic value.\n- Production readiness, mainnet, consensus, broad bridge finality, and arbitrary partition tolerance remain roadmap.\n"
-}
-
-fn report_status(input: &DemoReportInput<'_>) -> &'static str {
+pub(crate) fn report_status(input: &DemoReportInput<'_>) -> &'static str {
     if input.devnet_mode == "required" && input.devnet_settlement.is_none() {
         "NO-GO"
     } else {
