+++ crates/kamn-e2e-harness/src/mvp_demo/runner.rs
@@ -1,4 +1,4 @@
-use std::path::{Path, PathBuf};
+use std::path::Path;
 use std::time::{SystemTime, UNIX_EPOCH};
 
 use super::devnet_settlement::{
@@ -6,7 +6,8 @@ use super::devnet_settlement::{
 };
 use super::local_artifacts::create_demo_artifacts;
 use super::localhost_signed::{run_localhost_signed_demo, LocalhostSignedDemoInput};
-use super::report::{escape_json, render_report_json, render_report_markdown, DemoReportInput};
+use super::report::{escape_json, render_report_json, DemoReportInput};
+use super::report_writer::write_reports;
 use super::service_api_proof::{run_service_api_proofs, ServiceApiProofInput};
 use super::verify::verify_mvp_demo_report_json;
 
@@ -155,61 +156,5 @@ fn create_service_api_artifacts(
         websocket_log: run_dir
             .join("proof/service-api-websocket-output.txt")
             .as_path(),
-    })
-}
-
-fn write_reports(
-    output_root: &Path,
-    run_id: &str,
-    report_json: &str,
-    input: &DemoReportInput<'_>,
-) -> Result<(), String> {
-    let report_md = render_report_markdown(input);
-    write_report_pair(output_root.join(run_id), report_json, report_md.as_str())?;
-    refresh_latest(output_root, run_id, report_json, report_md.as_str())
-}
-
-fn write_report_pair(root: PathBuf, report_json: &str, report_md: &str) -> Result<(), String> {
-    let proof_dir = root.join("proof");
-    create_dir(&proof_dir)?;
-    write_file(proof_dir.join("report.json"), report_json)?;
-    write_file(proof_dir.join("report.md"), report_md)
-}
-
-fn refresh_latest(
-    output_root: &Path,
-    run_id: &str,
-    report_json: &str,
-    report_md: &str,
-) -> Result<(), String> {
-    let latest = output_root.join("latest");
-    remove_latest(&latest)?;
-    write_report_pair(latest.clone(), report_json, report_md)?;
-    write_file(latest.join("RUN_ID"), run_id)
-}
-
-fn remove_latest(latest: &Path) -> Result<(), String> {
-    if latest.exists() {
-        std::fs::remove_dir_all(latest)
-            .map_err(|error| format!("failed to remove previous latest demo: {error}"))?;
-    }
-    Ok(())
-}
-
-fn create_dir(path: &Path) -> Result<(), String> {
-    std::fs::create_dir_all(path).map_err(|error| {
-        format!(
-            "failed to create MVP demo directory {}: {error}",
-            path.display()
-        )
-    })
-}
-
-fn write_file(path: PathBuf, content: &str) -> Result<(), String> {
-    std::fs::write(&path, content).map_err(|error| {
-        format!(
-            "failed to write MVP demo artifact {}: {error}",
-            path.display()
-        )
     })
 }
