+++ crates/kamn-e2e-harness/src/mvp_demo/three_agent_claim.rs
@@ -9,38 +9,106 @@ pub(crate) fn three_agent_escrow_claim_json(
     let terms_digest = format!("terms-digest-{run_id}");
     let escrow_id = format!("escrow-three-agent-{run_id}");
     format!(
-        "{{\"id\":\"three_agent_escrow_verification\",\"label\":\"{}\",\"required\":true,\"status\":\"PASS\",\"summary\":\"Agent C verifies escrow settlement from restricted proof view\",\"transaction_id\":\"{}\",\"terms_digest\":\"{}\",\"agent_a_terms_digest\":\"{}\",\"agent_b_terms_digest\":\"{}\",\"verifier_terms_digest\":\"{}\",\"escrow_id\":\"{}\",\"agent_a_escrow_id\":\"{}\",\"agent_b_escrow_id\":\"{}\",\"verifier_escrow_id\":\"{}\",\"network\":\"{}\",\"rpc_url\":\"{}\",\"payer_pubkey\":\"{}\",\"recipient_pubkey\":\"{}\",\"lamports\":{},\"settlement_tx_signature\":\"{}\",\"agent_a_settlement_tx_signature\":\"{}\",\"agent_b_settlement_tx_signature\":\"{}\",\"verifier_settlement_tx_signature\":\"{}\",\"settlement_commitment\":\"{}\",\"agent_a_settlement_commitment\":\"{}\",\"agent_b_settlement_commitment\":\"{}\",\"verifier_settlement_commitment\":\"{}\",\"payer_balance_before\":{},\"payer_balance_after\":{},\"recipient_balance_before\":{},\"recipient_balance_after\":{},\"persisted_settlement_tx_signature\":\"{}\",\"amount_lamports\":{},\"agent_a_amount_lamports\":{},\"agent_b_amount_lamports\":{},\"verifier_amount_lamports\":{},\"agent_a_private_view_visible\":true,\"agent_b_private_view_visible\":true,\"verifier_private_view_visible\":false}}",
+        "{{{},{},{},{},{}}}",
+        claim_header(),
+        digest_fields(transaction_id.as_str(), terms_digest.as_str()),
+        escrow_fields(escrow_id.as_str()),
+        settlement_fields(evidence),
+        privacy_fields()
+    )
+}
+
+fn claim_header() -> String {
+    format!(
+        "\"id\":\"three_agent_escrow_verification\",\"label\":\"{}\",\"required\":true,\"status\":\"PASS\",\"summary\":\"Agent C verifies escrow settlement from restricted proof view\"",
         CLAIM_LABEL_DEVNET_BACKED,
-        escape_json(transaction_id.as_str()),
-        escape_json(terms_digest.as_str()),
-        escape_json(terms_digest.as_str()),
-        escape_json(terms_digest.as_str()),
-        escape_json(terms_digest.as_str()),
-        escape_json(escrow_id.as_str()),
-        escape_json(escrow_id.as_str()),
-        escape_json(escrow_id.as_str()),
-        escape_json(escrow_id.as_str()),
+    )
+}
+
+fn digest_fields(transaction_id: &str, terms_digest: &str) -> String {
+    format!(
+        "\"transaction_id\":\"{}\",\"terms_digest\":\"{}\",\"agent_a_terms_digest\":\"{}\",\"agent_b_terms_digest\":\"{}\",\"verifier_terms_digest\":\"{}\"",
+        escape_json(transaction_id),
+        escape_json(terms_digest),
+        escape_json(terms_digest),
+        escape_json(terms_digest),
+        escape_json(terms_digest),
+    )
+}
+
+fn escrow_fields(escrow_id: &str) -> String {
+    format!(
+        "\"escrow_id\":\"{}\",\"agent_a_escrow_id\":\"{}\",\"agent_b_escrow_id\":\"{}\",\"verifier_escrow_id\":\"{}\"",
+        escape_json(escrow_id),
+        escape_json(escrow_id),
+        escape_json(escrow_id),
+        escape_json(escrow_id),
+    )
+}
+
+fn settlement_fields(evidence: &DevnetSettlementEvidence) -> String {
+    [
+        network_fields(evidence),
+        signature_fields(evidence),
+        commitment_fields(evidence),
+        balance_fields(evidence),
+        amount_fields(evidence),
+    ]
+    .join(",")
+}
+
+fn network_fields(evidence: &DevnetSettlementEvidence) -> String {
+    format!(
+        "\"network\":\"{}\",\"rpc_url\":\"{}\",\"payer_pubkey\":\"{}\",\"recipient_pubkey\":\"{}\"",
         escape_json(evidence.network.as_str()),
         escape_json(evidence.rpc_url.as_str()),
         escape_json(evidence.payer_pubkey.as_str()),
         escape_json(evidence.recipient_pubkey.as_str()),
+    )
+}
+
+fn signature_fields(evidence: &DevnetSettlementEvidence) -> String {
+    format!(
+        "\"settlement_tx_signature\":\"{}\",\"agent_a_settlement_tx_signature\":\"{}\",\"agent_b_settlement_tx_signature\":\"{}\",\"verifier_settlement_tx_signature\":\"{}\"",
+        escape_json(evidence.settlement_tx_signature.as_str()),
+        escape_json(evidence.settlement_tx_signature.as_str()),
+        escape_json(evidence.settlement_tx_signature.as_str()),
+        escape_json(evidence.settlement_tx_signature.as_str()),
+    )
+}
+
+fn commitment_fields(evidence: &DevnetSettlementEvidence) -> String {
+    format!(
+        "\"settlement_commitment\":\"{}\",\"agent_a_settlement_commitment\":\"{}\",\"agent_b_settlement_commitment\":\"{}\",\"verifier_settlement_commitment\":\"{}\"",
+        escape_json(evidence.settlement_commitment.as_str()),
+        escape_json(evidence.settlement_commitment.as_str()),
+        escape_json(evidence.settlement_commitment.as_str()),
+        escape_json(evidence.settlement_commitment.as_str()),
+    )
+}
+
+fn balance_fields(evidence: &DevnetSettlementEvidence) -> String {
+    format!(
+        "\"lamports\":{},\"payer_balance_before\":{},\"payer_balance_after\":{},\"recipient_balance_before\":{},\"recipient_balance_after\":{},\"persisted_settlement_tx_signature\":\"{}\"",
         evidence.lamports,
-        escape_json(evidence.settlement_tx_signature.as_str()),
-        escape_json(evidence.settlement_tx_signature.as_str()),
-        escape_json(evidence.settlement_tx_signature.as_str()),
-        escape_json(evidence.settlement_tx_signature.as_str()),
-        escape_json(evidence.settlement_commitment.as_str()),
-        escape_json(evidence.settlement_commitment.as_str()),
-        escape_json(evidence.settlement_commitment.as_str()),
-        escape_json(evidence.settlement_commitment.as_str()),
         evidence.payer_balance_before,
         evidence.payer_balance_after,
         evidence.recipient_balance_before,
         evidence.recipient_balance_after,
         escape_json(evidence.persisted_settlement_tx_signature.as_str()),
+    )
+}
+
+fn amount_fields(evidence: &DevnetSettlementEvidence) -> String {
+    format!(
+        "\"amount_lamports\":{},\"agent_a_amount_lamports\":{},\"agent_b_amount_lamports\":{},\"verifier_amount_lamports\":{}",
         evidence.lamports,
         evidence.lamports,
         evidence.lamports,
         evidence.lamports
     )
+}
+
+fn privacy_fields() -> &'static str {
+    "\"agent_a_private_view_visible\":true,\"agent_b_private_view_visible\":true,\"verifier_private_view_visible\":false"
 }
